chore: rename to next-enterprise
