--- .github/workflows/deploy-expirator-service.yml
@@ -1,21 +1,0 @@
-name: Fly Deploy Expirator Service
-
-on:
-  push:
-    branches: [main]
-    paths:
-      - "next-enterprise/expirator-service/**"
-
-concurrency:
-  group: ${{ github.workflow }}
-  cancel-in-progress: true
-
-jobs:
-  expirator-service:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v6
-      - uses: superfly/flyctl-actions/setup-flyctl@fc53c09e1bc3be6f54706524e3b82c4f462f77be # v1
-      - run: flyctl deploy --remote-only ./next-enterprise/expirator-service
-        env:
-          FLY_API_TOKEN: ${{ secrets.FLY_API_TOKEN }}
