--- next-16-functions/next.config.ts
@@ -1,18 +1,0 @@
-import type {NextConfig} from 'next'
-
-const nextConfig: NextConfig = {
-  cacheComponents: true,
-  cacheLife: {
-    default: {
-      revalidate: 60 * 60 * 24 * 90,
-    },
-  },
-  logging: {
-    fetches: {
-      fullUrl: true,
-    },
-  },
-  reactCompiler: true,
-}
-
-export default nextConfig
