--- next-16-functions/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "next-16-functions",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "build": "next build",
-    "dev": "next dev",
-    "lint": "eslint",
-    "start": "next start",
-    "typegen:generate": "sanity typegen generate"
-  },
-  "dependencies": {
-    "@repo/generate-theme-colors": "workspace:*",
-    "@sanity/client": "^7.21.0",
-    "@vercel/speed-insights": "^1.3.1",
-    "groq": "^5.4.0",
-    "motion": "^12.26.2",
-    "next": "16.1.3",
-    "react": "^19.2.3",
-    "react-dom": "^19.2.3"
-  },
-  "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.18",
-    "@types/node": "^24.10.9",
-    "@types/react": "^19.2.8",
-    "@types/react-dom": "^19.2.3",
-    "babel-plugin-react-compiler": "^1.0.0",
-    "eslint": "^9.39.2",
-    "eslint-config-next": "16.1.2",
-    "eslint-plugin-react-hooks": "^7.0.1",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.18",
-    "typescript": "5.9.3"
-  }
-}
