--- next-16-functions/sanity.cli.ts
@@ -1,8 +1,0 @@
-import {defineCliConfig} from 'sanity/cli'
-
-export default defineCliConfig({
-  api: {projectId: 'hiomol4a', dataset: 'lcapi'},
-  typegen: {
-    schema: '../studio/schema.json',
-  },
-})
