--- next-16-functions/sanity.types.ts
@@ -1,246 +1,0 @@
-// Query TypeMap
-import '@sanity/client'
-
-/**
- * ---------------------------------------------------------------------------------
- * This file has been generated by Sanity TypeGen.
- * Command: `sanity typegen generate`
- *
- * Any modifications made directly to this file will be overwritten the next time
- * the TypeScript definitions are generated. Please make changes to the Sanity
- * schema definitions and/or GROQ queries if you need to update these types.
- *
- * For more information on how to use Sanity TypeGen, visit the official documentation:
- * https://www.sanity.io/docs/sanity-typegen
- * ---------------------------------------------------------------------------------
- */
-
-// Source: ../studio/schema.json
-export type Reaction = {
-  _id: string
-  _type: 'reaction'
-  _createdAt: string
-  _updatedAt: string
-  _rev: string
-  emoji?: string
-  reactions?: number
-}
-
-export type ReactionReference = {
-  _ref: string
-  _type: 'reference'
-  _weak?: boolean
-  [internalGroqTypeReferenceTo]?: 'reaction'
-}
-
-export type Demo = {
-  _id: string
-  _type: 'demo'
-  _createdAt: string
-  _updatedAt: string
-  _rev: string
-  title?: string
-  reactions?: Array<
-    {
-      _key: string
-    } & ReactionReference
-  >
-  slug?: Slug
-  url?: string
-}
-
-export type Slug = {
-  _type: 'slug'
-  current?: string
-  source?: string
-}
-
-export type Theme = {
-  _id: string
-  _type: 'theme'
-  _createdAt: string
-  _updatedAt: string
-  _rev: string
-  background?: string
-  text?: string
-}
-
-export type SanityImagePaletteSwatch = {
-  _type: 'sanity.imagePaletteSwatch'
-  background?: string
-  foreground?: string
-  population?: number
-  title?: string
-}
-
-export type SanityImagePalette = {
-  _type: 'sanity.imagePalette'
-  darkMuted?: SanityImagePaletteSwatch
-  lightVibrant?: SanityImagePaletteSwatch
-  darkVibrant?: SanityImagePaletteSwatch
-  vibrant?: SanityImagePaletteSwatch
-  dominant?: SanityImagePaletteSwatch
-  lightMuted?: SanityImagePaletteSwatch
-  muted?: SanityImagePaletteSwatch
-}
-
-export type SanityImageDimensions = {
-  _type: 'sanity.imageDimensions'
-  height?: number
-  width?: number
-  aspectRatio?: number
-}
-
-export type SanityImageMetadata = {
-  _type: 'sanity.imageMetadata'
-  location?: Geopoint
-  dimensions?: SanityImageDimensions
-  palette?: SanityImagePalette
-  lqip?: string
-  blurHash?: string
-  hasAlpha?: boolean
-  isOpaque?: boolean
-}
-
-export type SanityImageHotspot = {
-  _type: 'sanity.imageHotspot'
-  x?: number
-  y?: number
-  height?: number
-  width?: number
-}
-
-export type SanityImageCrop = {
-  _type: 'sanity.imageCrop'
-  top?: number
-  bottom?: number
-  left?: number
-  right?: number
-}
-
-export type SanityFileAsset = {
-  _id: string
-  _type: 'sanity.fileAsset'
-  _createdAt: string
-  _updatedAt: string
-  _rev: string
-  originalFilename?: string
-  label?: string
-  title?: string
-  description?: string
-  altText?: string
-  sha1hash?: string
-  extension?: string
-  mimeType?: string
-  size?: number
-  assetId?: string
-  uploadId?: string
-  path?: string
-  url?: string
-  source?: SanityAssetSourceData
-}
-
-export type SanityAssetSourceData = {
-  _type: 'sanity.assetSourceData'
-  name?: string
-  id?: string
-  url?: string
-}
-
-export type SanityImageAsset = {
-  _id: string
-  _type: 'sanity.imageAsset'
-  _createdAt: string
-  _updatedAt: string
-  _rev: string
-  originalFilename?: string
-  label?: string
-  title?: string
-  description?: string
-  altText?: string
-  sha1hash?: string
-  extension?: string
-  mimeType?: string
-  size?: number
-  assetId?: string
-  uploadId?: string
-  path?: string
-  url?: string
-  metadata?: SanityImageMetadata
-  source?: SanityAssetSourceData
-}
-
-export type Geopoint = {
-  _type: 'geopoint'
-  lat?: number
-  lng?: number
-  alt?: number
-}
-
-export type AllSanitySchemaTypes =
-  | Reaction
-  | ReactionReference
-  | Demo
-  | Slug
-  | Theme
-  | SanityImagePaletteSwatch
-  | SanityImagePalette
-  | SanityImageDimensions
-  | SanityImageMetadata
-  | SanityImageHotspot
-  | SanityImageCrop
-  | SanityFileAsset
-  | SanityAssetSourceData
-  | SanityImageAsset
-  | Geopoint
-
-export declare const internalGroqTypeReferenceTo: unique symbol
-
-// Source: src/app/Reactions.tsx
-// Variable: REACTION_QUERY
-// Query: *[_type == "reaction" && _id == $id][0]{emoji,reactions}
-export type REACTION_QUERY_RESULT = {
-  emoji: string | null
-  reactions: number | null
-} | null
-
-// Source: src/app/dynamic/page.tsx
-// Variable: DYNAMIC_DEMO_QUERY
-// Query: *[_type == "demo" && slug.current == $slug][0]{title}
-export type DYNAMIC_DEMO_QUERY_RESULT = {
-  title: string | null
-} | null
-
-// Source: src/app/layout.tsx
-// Variable: THEME_QUERY
-// Query: *[_id == "theme"][0]{background,text}
-export type THEME_QUERY_RESULT =
-  | {
-      background: null
-      text: null
-    }
-  | {
-      background: string | null
-      text: string | null
-    }
-  | null
-
-// Source: src/app/page.tsx
-// Variable: DEMO_QUERY
-// Query: *[_type == "demo" && slug.current == $slug][0]{title,reactions[0..4]{_key,_ref}}
-export type DEMO_QUERY_RESULT = {
-  title: string | null
-  reactions: Array<{
-    _key: string
-    _ref: string
-  }> | null
-} | null
-
-declare module '@sanity/client' {
-  interface SanityQueries {
-    '*[_type == "reaction" && _id == $id][0]{emoji,reactions}': REACTION_QUERY_RESULT
-    '*[_type == "demo" && slug.current == $slug][0]{title}': DYNAMIC_DEMO_QUERY_RESULT
-    '*[_id == "theme"][0]{background,text}': THEME_QUERY_RESULT
-    '*[_type == "demo" && slug.current == $slug][0]{title,reactions[0..4]{_key,_ref}}': DEMO_QUERY_RESULT
-  }
-}
