--- next-16-functions/src/app/globals.css
@@ -1,25 +1,0 @@
-@import 'tailwindcss';
-
-@theme {
-  --background-color-theme: var(--theme-background, #fff);
-  --background-color-theme-button: var(--theme-text, #fff);
-
-  --text-color-theme: var(--theme-text, #000);
-  --text-color-theme-button: var(--theme-background, #fff);
-
-  --ring-color-theme: var(--theme-text, #000);
-
-  --ring-offset-color-theme: var(--theme-background, #fff);
-}
-
-@property --theme-text {
-  syntax: '<color>';
-  inherits: true;
-  initial-value: #000;
-}
-
-@property --theme-background {
-  syntax: '<color>';
-  inherits: true;
-  initial-value: #fff;
-}
