--- next-16-functions/src/sanity/client.ts
@@ -1,8 +1,0 @@
-import {createClient} from '@sanity/client'
-
-export const client = createClient({
-  projectId: 'hiomol4a',
-  dataset: 'lcapi',
-  apiVersion: '2025-10-23',
-  useCdn: true,
-})
