--- next-enterprise/.env.local.example
@@ -1,1 +1,0 @@
-SECRET='your-secret'
