+++ next-enterprise/.gitignore
@@ -31,7 +31,6 @@ yarn-error.log*
 
 # env files (can opt-in for committing if needed)
 .env*
-!.env.local.example
 
 # vercel
 .vercel
