--- next-enterprise/expirator-service/.env.local.example
@@ -1,2 +1,0 @@
-SECRET='your-secret'
-TOKEN='sk...
