--- next-enterprise/expirator-service/Dockerfile
@@ -1,37 +1,0 @@
-# syntax = docker/dockerfile:1
-
-FROM oven/bun:1.3.6-slim AS base
-
-LABEL fly_launch_runtime="Bun"
-
-# Bun app lives here
-WORKDIR /app
-
-# Set production environment
-ENV NODE_ENV="production"
-
-
-# Throw-away build stage to reduce size of final image
-FROM base AS build
-
-# Install packages needed to build node modules
-RUN apt-get update -qq && \
-    apt-get install --no-install-recommends -y build-essential pkg-config python-is-python3
-
-# Install node modules
-COPY bun.lockb package.json ./
-RUN bun install --ci
-
-# Copy application code
-COPY . .
-
-
-# Final stage for app image
-FROM base
-
-# Copy built application
-COPY --from=build /app /app
-
-# Start the server by default, this can be overwritten at runtime
-EXPOSE 3000
-CMD [ "bun", "run", "start" ]
