--- next-enterprise/expirator-service/fly.toml
@@ -1,18 +1,0 @@
-# fly.toml app configuration file generated for expirator-service on 2025-02-18T23:02:00+01:00
-#
-# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
-#
-
-app = 'expirator-service'
-primary_region = 'cdg'
-
-[build]
-
-[[vm]]
-  size = 'shared-cpu-1x'
-  memory = '256mb'
-  count = 1
-
-[scale]
-  min_machines_count = 1
-  max_machines_count = 1
