--- next-enterprise/expirator-service/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "expirator-service",
-  "type": "module",
-  "module": "index.ts",
-  "scripts": {
-    "lint": "tsc",
-    "start": "bun run index.ts"
-  },
-  "dependencies": {
-    "@sanity/client": "^7.14.1"
-  },
-  "devDependencies": {
-    "@flydotio/dockerfile": "^0.7.10",
-    "@types/bun": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5.0.0"
-  }
-}
