+++ next-enterprise/package.json
@@ -11,18 +11,18 @@
   },
   "dependencies": {
     "@repo/generate-theme-colors": "workspace:*",
-    "@sanity/client": "^7.14.1",
+    "@sanity/client": "^7.21.0",
     "@vercel/speed-insights": "^1.3.1",
-    "groq": "^5.7.0",
-    "motion": "^12.30.1",
-    "next": "16.1.5",
-    "react": "^19.2.4",
-    "react-dom": "^19.2.4"
+    "groq": "^5.4.0",
+    "motion": "^12.26.2",
+    "next": "16.1.3",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.1.18",
-    "@types/node": "^24.10.10",
-    "@types/react": "^19.2.10",
+    "@types/node": "^24.10.9",
+    "@types/react": "^19.2.8",
     "@types/react-dom": "^19.2.3",
     "babel-plugin-react-compiler": "^1.0.0",
     "eslint": "^9.39.2",
