+++ next-enterprise/sanity.types.ts
@@ -196,19 +196,19 @@ export type AllSanitySchemaTypes =
 
 export declare const internalGroqTypeReferenceTo: unique symbol
 
-// Source: src/app/api/reaction/[id]/route.ts
-// Variable: REACTION_QUERY_STATIC_PARAMS
-// Query: *[_type == "demo" && slug.current == $slug][0].reactions[0..4]{"id": _ref}
-export type REACTION_QUERY_STATIC_PARAMS_RESULT = Array<{
-  id: string
-}> | null
-
-// Source: src/app/api/reaction/[id]/shared.ts
-// Variable: API_REACTION_QUERY
+// Source: src/app/Reactions.tsx
+// Variable: REACTION_QUERY
 // Query: *[_type == "reaction" && _id == $id][0]{emoji,reactions}
-export type API_REACTION_QUERY_RESULT = {
+export type REACTION_QUERY_RESULT = {
   emoji: string | null
   reactions: number | null
+} | null
+
+// Source: src/app/dynamic/page.tsx
+// Variable: DYNAMIC_DEMO_QUERY
+// Query: *[_type == "demo" && slug.current == $slug][0]{title}
+export type DYNAMIC_DEMO_QUERY_RESULT = {
+  title: string | null
 } | null
 
 // Source: src/app/layout.tsx
@@ -238,8 +238,8 @@ export type DEMO_QUERY_RESULT = {
 
 declare module '@sanity/client' {
   interface SanityQueries {
-    '*[_type == "demo" && slug.current == $slug][0].reactions[0..4]{"id": _ref}': REACTION_QUERY_STATIC_PARAMS_RESULT
-    '*[_type == "reaction" && _id == $id][0]{emoji,reactions}': API_REACTION_QUERY_RESULT
+    '*[_type == "reaction" && _id == $id][0]{emoji,reactions}': REACTION_QUERY_RESULT
+    '*[_type == "demo" && slug.current == $slug][0]{title}': DYNAMIC_DEMO_QUERY_RESULT
     '*[_id == "theme"][0]{background,text}': THEME_QUERY_RESULT
     '*[_type == "demo" && slug.current == $slug][0]{title,reactions[0..4]{_key,_ref}}': DEMO_QUERY_RESULT
   }
