+++ next-enterprise/src/app/actions.ts
@@ -1,6 +1,14 @@
 'use server'
 
-import {updateTag} from 'next/cache'
+import type {SyncTag} from '@sanity/client'
+import {refresh, updateTag} from 'next/cache'
+
+export async function updateTags(tags: SyncTag[]) {
+  for (const tag of tags) {
+    updateTag(tag)
+  }
+  console.log(`<SanityLive /> updated tags: ${tags.join(', ')}`)
+}
 
 export async function randomColorTheme(background: string, text: string) {
   const formData = new FormData()
@@ -11,4 +19,8 @@ export async function randomColorTheme(background: string, text: string) {
     body: formData,
   })
   updateTag('theme')
+}
+
+export async function liveRefresh() {
+  refresh()
 }
