+++ next-enterprise/src/app/dynamic/UserAgent.tsx
