+++ next-enterprise/src/sanity/client.ts
@@ -3,6 +3,6 @@ import {createClient} from '@sanity/client'
 export const client = createClient({
   projectId: 'hiomol4a',
   dataset: 'lcapi',
-  apiVersion: '2025-10-21',
+  apiVersion: '2025-10-23',
   useCdn: true,
 })
