+++ pnpm-workspace.yaml
@@ -1,7 +1,6 @@
 packages:
   - '*'
   - 'packages/*'
-  - 'next-enterprise/expirator-service'
   - '!node_modules'
 
 publicHoistPattern:
