+++ studio/functions/cache-invalidate/index.ts
@@ -1,7 +1,7 @@
 import {syncTagInvalidateEventHandler} from '@sanity/functions'
 
 const REVALIDATE_URL =
-  'https://lcapi-examples-next-16-functions.sanity.dev/api/revalidate'
+  'https://lcapi-examples-next-enterprise.sanity.dev/api/revalidate'
 
 export const handler = syncTagInvalidateEventHandler(async ({event, done}) => {
   const {syncTags} = event.data
